Rename section_toggle controller to form_section_toggle
Scope the controller name to forms to avoid ambiguity with other
section-related controllers.
