+++ app/frontend/javascript/controllers/form_section_toggle_controller.js
