+++ app/frontend/javascript/controllers/index.js
@@ -126,8 +126,8 @@ application.register("searchable-select", SearchableSelectController)
 import SectionFilterController from "./section_filter_controller"
 application.register("section-filter", SectionFilterController)
 
-import SectionToggleController from "./section_toggle_controller"
-application.register("section-toggle", SectionToggleController)
+import FormSectionToggleController from "./form_section_toggle_controller"
+application.register("form-section-toggle", FormSectionToggleController)
 
 import ShareUrlController from "./share_url_controller"
 application.register("share-url", ShareUrlController)
